[IMP] add files in report folder and chnges in wizard.
