+++ report/__init__.py
@@ -1,3 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from . import church_attendance_report
+from . import tithe_report
+from . import pledges_report
+from . import offering_report
+from . import donation_report
