+++ wizard/Pledge.py
@@ -1,39 +1,9 @@
 # -*- coding: utf-8 -*-
 """."""
 
-from odoo import api, fields, models
+from odoo import fields, models
 import datetime
 from odoo.exceptions import MissingError
-
-
-class PledgesReport(models.AbstractModel):
-    """PledgesReport."""
-
-    _name = 'report.ng_church.church_pledges_report'
-    _description = "Report NG Church Church Pledges Report"
-
-    # def reports_presenter(self, model):
-    #     """Mutate the state of the original report(s)."""
-    #     return model
-
-    @api.model
-    def _get_report_values(self, docids, data=None):
-        docs = self.env['ng_church.pledge'].browse(
-            data.get('docids'))
-        pledge_lines = self.env['ng_church.pledge_line'].browse(
-            data.get('pledge_lines'))
-        return {
-            'doc_ids': docs.ids,
-            'doc_model': 'ng_church.pledge',
-            'docs': docs,
-            'pledge_lines': pledge_lines,
-        }
-        # docs = self.env['ng_church.pledge'].browse(docids)
-        # return {
-        #     'doc_ids': docs.ids,
-        #     'doc_model': 'ng_church.pledge',
-        #     'docs': self.env['ng_church.pledge'].browse(docids),
-        # }
 
 
 class ChurchPledgeReport(models.TransientModel):
